+++ NEWS.md
@@ -1,5 +1,8 @@
 # xaringanExtra (development version)
 
+* **scribble** can now be disabled on an individual slide by adding
+  `class: no-scribble` to the slide (@mattwarkentin #166).
+
 # xaringanExtra 0.6.0 (2022-06-07)
 
 * First release available on CRAN! Read about it at [garrickadenbuie.com/blog/xaringanextra-v0.6.0/](https://www.garrickadenbuie.com/blog/xaringanextra-v0.6.0/).